Adicionando gate authorize

+++ app/Http/Controllers/AssistantController.php
@@ -5,13 +5,16 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rules;
 
 class AssistantController extends Controller
 {
     //Gerencia os assistentes
-    public function index(){
+    public function index()
+    {
+        Gate::authorize('manage-assistants');
 
         // Busca todos os usuários que têm role_id igual a 3 (assistentes)
         $assistants = User::where('role_id', 3)->get();
@@ -22,12 +25,15 @@ class AssistantController extends Controller
     //View form create assistant
     public function create()
     {
+        Gate::authorize('manage-assistants');
         return view('assistant.create');
     }
 
     //Cria usuario assistant
     public function store(Request $request): RedirectResponse
     {
+        Gate::authorize('manage-assistants');
+
         $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'lowercase', 'email', 'max:255', 'unique:'.User::class],
@@ -47,7 +53,9 @@ class AssistantController extends Controller
     }
 
     //Deleta usuario assistente
-    public function destroy($id){
+    public function destroy($id)
+    {
+        Gate::authorize('manage-assistants');
 
         User::destroy($id);
